+++ internal/util/reports.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"context"
 	"crypto/tls"
 	"fmt"
 	"net"
@@ -9,6 +10,8 @@ import (
 	"time"
 
 	"github.com/sirupsen/logrus"
+
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/meshdetect"
 )
 
 var (
@@ -42,6 +45,9 @@ type Reporter struct {
 	serializedInfo string
 
 	Logger logrus.FieldLogger
+
+	MeshDetectionEnabled bool
+	MeshDetector         *meshdetect.Detector
 }
 
 func (r *Reporter) once() {
@@ -91,6 +97,25 @@ func (r *Reporter) sendPing(uptime int) {
 func (r *Reporter) send(signal string, uptime int) {
 	message := "<14>signal=" + signal + ";uptime=" +
 		strconv.Itoa(uptime) + ";" + r.serializedInfo
+	// run mesh detection if enabled.
+	if r.MeshDetectionEnabled {
+		meshMessage, err := r.getMeshMessages(context.Background())
+		if err != nil {
+			// log the error if mesh detection fails,
+			// but still send the messages without mesh detection results.
+			r.Logger.Debugf("failed to run mesh detection: %w", err)
+
+		} else {
+			// append results from mesh detection to reported message if returned any.
+			if meshMessage != "" {
+				if !strings.HasSuffix(message, ";") {
+					message = message + ";"
+				}
+				message = message + meshMessage
+			}
+		}
+	}
+
 	conn, err := tls.DialWithDialer(&dialer, "tcp", net.JoinHostPort(reportsHost,
 		strconv.FormatUint(uint64(reportsPort), 10)), &tlsConf)
 	if err != nil {
@@ -107,4 +132,25 @@ func (r *Reporter) send(signal string, uptime int) {
 	if err != nil {
 		r.Logger.Debugf("failed to send report: %s", err)
 	}
+}
+
+// getMeshMessages returns serialized messages of results of mesh detection.
+func (r *Reporter) getMeshMessages(ctx context.Context) (string, error) {
+	if r.MeshDetector == nil {
+		return "", fmt.Errorf("no mesh detector")
+	}
+
+	meshMessages := []string{}
+	deploymentResults := r.MeshDetector.DetectMeshDeployment(ctx)
+	meshMessages = append(meshMessages, serializeMeshDeploymentResults(deploymentResults))
+
+	runUnderResults := r.MeshDetector.DetectRunUnder(ctx)
+	meshMessages = append(meshMessages, serializeMeshRunUnderResults(runUnderResults))
+
+	serviceDistributionResults, detectErr := r.MeshDetector.DetectServiceDistribution(ctx)
+	if detectErr != nil {
+		return "", fmt.Errorf("failed to detect service distribution under meshes: %w", detectErr)
+	}
+	meshMessages = append(meshMessages, serializeMeshServiceDistribution(serviceDistributionResults))
+	return strings.Join(meshMessages, ";"), nil
 }